requirements.txt
fastapi
uvicorn
openai
python-dotenv
supabase
pdfplumber
python-docx
pydantic==2.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9 
requests
